Folding examples: remove duplicated checker implementation
